🏗️ Architect: Modularized weather analysis into a package.
Transformed the monolithic `apts/observations/weather.py` into a modular package `apts/observations/weather/`. This refactoring improves code organization by separating core MixIn logic from analysis algorithms and constants, while significantly reducing the cyclomatic complexity of weather suitability checks.

Key changes:
- Created `apts/observations/weather/` package.
- Decomposed `WeatherAnalysisMixIn` into `base.py` (orchestration), `engine.py` (vectorized analysis functions), and `constants.py` (default values).
- Maintained backward compatibility via `__init__.py` re-exports.
- Updated unit tests to correctly mock `get_moon_illumination` within the new module structure.
