+++ apts/observations/weather/constants.py
@@ -1,0 +1,13 @@
+DEFAULT_WEATHER_VALUES = {
+    "fog": 0,
+    "aurora": 0,
+    "precipIntensity": 0,
+    "precipProbability": 0,
+    "cloudCover": 0,
+    "windSpeed": 0,
+    "temperature": 20,
+    "visibility": 20,
+    "moonIllumination": 0,
+    "seeing": 1.5,
+    "sqm": 21.0,
+}
